Fixes #221: src/families folder
