+++ .prettierignore
@@ -1,2 +1,3 @@
 package.json
 src/load
+src/generated
